// data/encounters/enums.ts
export enum EncounterId {
  A1 = "A1",
}
export enum EncounterPhase {
  ENCOUNTER_START = "ENCOUNTER_START",
  PLAYER_PREP = "PLAYER_PREP",
  PLAYER_PHASE = "PLAYER_PHASE",
  NEW_ROUND = "NEW_ROUND",
  ENEMY_PHASE = "ENEMY_PHASE",
  ENCOUNTER_OVER = "ENCOUNTER_OVER",
}

export enum DrawPhase {
  INACTIVE = "INACTIVE",
  DRAW = "DRAW",
  SHUFFLE = "SHUFFLE",
  DISCARD = "DISCARD",
}